feat: add wikilink handler interface
This is a huge commit that also changes a bit of the code structure.
Arcs are now a part of the `vault` module, and wikilinks get a module
for themselves. The commit has a rewrite of how wikilinks are handled,
which allows users of the API to add their own handlers to get custom
functionality.

For now, the wikilink to CommonMark link conversion is crudely
implemented as a wikilink handler. It is used as a default in
VaultBuilder.
