+++ src/lib.rs
@@ -2,7 +2,7 @@ extern crate pest;
 #[macro_use]
 extern crate pest_derive;
 
-mod arc;
+mod wikilink;
 pub mod error;
 mod hash;
 mod note;
